api_service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
aiohttp==3.10.0
pillow==11.0.0
python-multipart==0.0.12
